32-bit platforms:
src/net.c: In function '_sendPing':
src/net.c:349: warning: format '%lu' expects type 'long unsigned int', but argument 6 has type 'unsigned int'

+++ src/net.c
@@ -346,7 +346,7 @@ static boolean_t _sendPing(const char *hostname, int socket, struct addrinfo *ad
                         break;
         }
         if (out_len > sizeof(buf)) {
-                LogError("Ping request for %s %d/%d failed -- too large (%d vs. maximum %lu bytes)\n", hostname, retry, maxretries, size, sizeof(buf) - header_len);
+                LogError("Ping request for %s %d/%d failed -- too large (%d vs. maximum %lu bytes)\n", hostname, retry, maxretries, size, (unsigned long)(sizeof(buf) - header_len));
                 return false;
         }
         ssize_t n;
